requirements.txt
Flask
Werkzeug
python-docx
mammoth
htmldiff2
Genshi
pypandoc
pandoc
gunicorn
